Active display: Enable proximity sensor only when screen is off
This should ease people worried about cpu usage of the proximity
sensor while the screen is on.  The sensor listener will only be
registered when the screen goes off, and both active display and
pocket mode are enabled.  Once the screen turns on this listener
will be unregistered so no further updates are received when the
sensor values change.

Change-Id: I70979b5931815adb74bd8bc42e32886a6d829a12

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/policy/activedisplay/ActiveDisplayView.java