Keyguard: Distinguish the TextView id of pinEntry
Both KeyguardSimPinView and KeyguardPinView have the same id of pinEntry
for their TextView of the password. When one of them is created in the
mSecurityViewContainer, the other one will get the wrong TextView of the
same id pinEntry.

Change the id pinEntry in KeyguardSimPinView to simPinEntry.

This fixes CYAN-3822,CYAN-3887

CRs-Fixed: 661955

Change-Id: Ia75d1d1f74530e6671d534000f9ce49cb1f814c4

Conflicts:
	packages/Keyguard/res/layout/msim_keyguard_sim_pin_view.xml