keyguard: refresh carrier info on subId changes
Currently, the SPN_STRINGS_UPDATED_ACTION intent receives the phoneId and subId, but sometimes
during boot the SubscriptionManager isn't be updated yet, which lead to SPN and PLMN were mapped
with an invalid id in the KeyguardUpdateMonitor, and CarrierText not being updated, although
in the phone implementation the data is correct.
This change updates the carrier info also on SUBID changes, so in case the subid were
updated after SPN_STRINGS_UPDATED_ACTION, all SPN and PLMN can be mapped properly to the
correct SUBID and CarrierText update with the right values.

Change-Id: Ia4243ee8781a31942cb4370489df17f9c970e41f
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>